Continued Refining Weapons, Added Armor, Began Using PlayerStats
Refined weapons in general. Damage, knockback, and attack time now equal
headSize * handleLength (refine this later?). The constructor for
weapons now takes only the floats handle length and head size along with
the weapon type. Renamed all "dmg" variables to have "damage."
Added an "Armor" class, which PlayerStats has an instance of. The
modifiers for stats in this class are applied to the default variables
to get the current value (see pseudocode for later plans).
Began using PlayerStats (and, by extension, armor) in that the
PlayerMove class now uses speed and jumpSpeed from the PlayerStats
class.