Customizable Armor! WHOO!
Armor is now customizable. Players have a pool of 10 points; they can
edit 10 stats to be between .5 and 1.5x normal. Some of those stats are
not yet implemented, but they are editable, and the ones that DO have an
effect are pretty cool (whoo! MOVESPEED!). You can access the armors
through the cube on the left through the interact button ('e' by
default).