Spells Sorta-Working!
So... Basically, you now have a spell control on your character, which
has a spell (data about firing a spell). The control can fire off a
projectile and attach the spell to it. That spell also has an effect,
which is applied to an AI if the projectile hits it. Currently the spell
is unmodifiable and can only be a projectile; the effect is 25 damage
(times armor resistance) and stops enemy regen for 3 seconds.