Tweaked Weapon Scaling
Placement of the "blades" of the three different weapons now places the
blade on the handle tip (and then offsets it back "into" the handle for
hammers/axes). It does this by adding half of the handle scale to the
placement, then subtracting half of the blade length, then offsetting
it. Also tweaked the swords' default scaling/placement.