Added Voxelmetric, Removed Torch/Shield Stuff
- Added the Voxelmetric block system to the game (used master branch)
- Removed currently unused shield and torch code. It may/will be added
back to the game later.
- Moved spawn point and customizer stations to be in nice places in the
world

This may be the last commit for a couple weeks, I'm really busy and
don't currently have the drive to keep programming. I didn't expect to
get as far as I have, so I don't have a clear idea of what to do next.