Added PlayerMode Script
PlayerMode is used to switch the player between "Build" and "Weapon"
modes. In Weapon mode, you can swing your weapon and fire spells, and
see them on screen. Build mode is currently not implemented, but in the
future, you will be able to place and remove blocks in this mode (duh).
You can switch between the modes using 'b.'

Also, changed initial jump speed from 8 to 6.5, which is a height change
from roughly three blocks to roughly two.