Beginning Game Logic Stuff
Separated Player and AI health. When AIs die, the game object and
healthbar object are both destroyed. Made health for all characters
reset to max on startup. Created a GameState script and GameLogic object
to keep track of whether enemies should be spawning. Created a spawner
script and object to spawn enemies. Added Interactor support for an
attack start object.

Enemies still do nothing, and even if they could, there is nothing for
them to attack.